backend/requirements-windows.txt
# Requirements alternativo para Windows - instala dependências uma por uma
# Execute: pip install -r requirements-windows.txt

Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8

# Instale psycopg2-binary separadamente se necessário:
# pip install psycopg2-binary

# Instale Pillow separadamente (versão mais recente com wheels):
# pip install Pillow







